# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MacroVanta API
  description: |
    REST API for the MacroVanta app. Handles user authentication with JWT,
    food entry logging, AI-assisted extraction, USDA food search, daily macro
    and micronutrient goals, and nutrition summaries.

    Authentication: obtain a token from `POST /api/auth/login`, then click
    Authorize and paste the token. SwaggerHub adds the `Bearer ` prefix when
    using the bearer auth helper.

    Email verification: accounts must verify their email before logging in.
    Unverified accounts receive a 403 from the login endpoint.

    Dates: all dates are plain strings in `YYYY-MM-DD` format. Summary endpoint
    defaults use the server's current UTC date.

    Development tokens: when the server is not running in production,
    token-generating responses include a `developmentToken` field. Registration
    includes it after creating a verification token. Resend verification includes
    it only when an existing unverified account receives a new verification
    token. Forgot password includes it only when an existing account receives a
    reset token.
  version: 1.1.0
  contact:
    name: Anthony Mainolfi
servers:
  - url: https://dylanwexler.com
    description: Deployed course demonstration server
  - url: http://localhost:5000
    description: Local development server
...
paths:
...
  /api/foods:
    post:
      tags: [Foods]
      summary: Create a food entry
      operationId: createFood
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodEntryInput"
      responses:
        "201":
          description: Food entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodEntryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      tags: [Foods]
      summary: List food entries
      description: Returns the user's entries, sorted by date descending and then creation time descending. Optionally filtered to a single date.
      operationId: getFoods
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DateQuery"
      responses:
        "200":
          description: Food entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodEntryListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/foods/search:
    get:
      tags: [Foods]
      summary: Search food entries by name
      description: Case-insensitive substring match on food name. Returns at most 50 results.
      operationId: searchFoods
      security:
        - bearerAuth: []
      parameters:
        - name: query
          in: query
          required: true
          description: Search text matched anywhere in the food name
          schema:
            type: string
          example: chicken
        - $ref: "#/components/parameters/DateQuery"
      responses:
        "200":
          description: Matching food entries, max 50
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodEntryListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/foods/ai-log:
    post:
      tags: [Foods]
      summary: Log consumed food from plain language
      description: |
        Uses a narrowly scoped LLM parser with strict structured output to
        estimate nutrition for food or drinks the authenticated user consumed.
        Input is limited to 500 characters and eight food items. The server
        rejects non-food content and instruction-like input, validates every
        returned value, labels entries as AI estimates, and rate limits each
        user. Estimates should be reviewed and edited when precision matters.
      operationId: logFoodWithAi
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiFoodLogRequest"
      responses:
        "201":
          description: One or more estimated food entries saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiFoodLogResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: Text did not describe food already consumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Per-user AI request limit reached
          headers:
            Retry-After:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: AI provider unavailable or returned invalid data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: OpenAI API key is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/foods/recent:
    get:
      tags: [Foods]
      summary: Get recent unique foods
      description: |
        Looks at the user's 50 most recent entries and returns up to 10 unique
        foods, de-duplicated by name, serving size, and macros. Useful for quick
        re-logging.
      operationId: getRecentFoods
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Recent unique food entries, max 10
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodEntryListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/foods/quick-add/{id}:
    post:
      tags: [Foods]
      summary: Quick-add a copy of an existing entry
      description: |
        Creates a new food entry by copying an existing one, using its name,
        serving size, and macros on the supplied date. Meal type defaults to the
        source entry's meal type unless overridden.
      operationId: quickAddFood
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FoodEntryId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuickAddFoodRequest"
      responses:
        "201":
          description: New food entry created from the source entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodEntryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/FoodNotFound"
  /api/foods/{id}:
    get:
      tags: [Foods]
      summary: Get a food entry by id
      operationId: getFoodById
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FoodEntryId"
      responses:
        "200":
          description: Food entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodEntryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/FoodNotFound"
    put:
      tags: [Foods]
      summary: Update a food entry
      description: Partial update. Send any subset of fields, but at least one field is required.
      operationId: updateFood
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FoodEntryId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodEntryUpdate"
      responses:
        "200":
          description: Updated food entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodEntryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/FoodNotFound"
    delete:
      tags: [Foods]
      summary: Delete a food entry
      operationId: deleteFood
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FoodEntryId"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
              example:
                message: Food entry deleted successfully.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/FoodNotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT returned by `POST /api/auth/login`
  parameters:
    DateQuery:
      name: date
      in: query
      required: false
      description: Filter by date, in YYYY-MM-DD format. Summary endpoints default to the current UTC date.
      schema:
        type: string
        pattern: '^\d{4}-\d{2}-\d{2}$'
      example: "2026-07-07"
    FoodEntryId:
      name: id
      in: path
      required: true
      description: Food entry id, stored as a MongoDB ObjectId
      schema:
        type: string
      example: 665f1c2ab8e4a91d2c3f0a11
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            invalidDate:
              value:
                message: date must use YYYY-MM-DD format.
            validation:
              value:
                message: At least one field is required.
            weakPassword:
              value:
                message: Password must be at least 8 characters and include uppercase, lowercase, number, and special character.
    Unauthorized:
      description: Missing, invalid, or expired authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            missing:
              value:
                message: Authentication token required.
            invalid:
              value:
                message: Invalid authentication token.
            expired:
              value:
                message: Invalid or expired authentication token.
    FoodNotFound:
      description: Food entry not found or not owned by this user
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            message: Food entry not found.
  schemas:
...
    MealType:
      type: string
      enum: [Breakfast, Lunch, Dinner, Snack]
      example: Lunch
    DateString:
      type: string
      pattern: '^\d{4}-\d{2}-\d{2}$'
      example: "2026-07-07"
    NonNegativeNumberInput:
      description: Backend coerces numeric strings to numbers; sending JSON numbers is recommended.
      oneOf:
        - type: number
          minimum: 0
        - type: string
          pattern: '^\d+(\.\d+)?$'
      example: 2400
...
    FoodEntry:
      type: object
      required: [id, foodName, servingSize, mealType, calories, protein, carbs, fat, saturatedFat, transFat, fiber, sodium, potassium, calcium, iron, vitaminC, vitaminD, source, confidence, date, createdAt, updatedAt]
      properties:
        id:
          type: string
          example: 665f1c2ab8e4a91d2c3f0a11
        foodName:
          type: string
          example: Grilled Chicken Breast
        servingSize:
          type: string
          example: 6 oz
        mealType:
          $ref: "#/components/schemas/MealType"
        calories:
          type: number
          example: 280
        protein:
          type: number
          example: 52
        carbs:
          type: number
          example: 0
        fat:
          type: number
          example: 6
        saturatedFat:
          type: number
          example: 1.5
        transFat:
          type: number
          example: 0
        fiber:
          type: number
          example: 0
        sodium:
          type: number
          example: 130
        potassium:
          type: number
          example: 450
        calcium:
          type: number
          example: 20
        iron:
          type: number
          example: 1
        vitaminC:
          type: number
          example: 0
        vitaminD:
          type: number
          example: 0
        source:
          type: string
          enum: [manual, ai, usda]
          example: manual
        confidence:
          type: string
          enum: [low, medium, high]
          nullable: true
          example: null
        date:
          $ref: "#/components/schemas/DateString"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FoodEntryInput:
      type: object
      required: [foodName, servingSize, mealType, calories, protein, carbs, saturatedFat, transFat, sodium, date]
      properties:
        foodName:
          type: string
          minLength: 1
          maxLength: 120
          example: Grilled Chicken Breast
        servingSize:
          type: string
          minLength: 1
          maxLength: 80
          example: 6 oz
        mealType:
          $ref: "#/components/schemas/MealType"
        calories:
          $ref: "#/components/schemas/NonNegativeNumberInput"
        protein:
          allOf:
            - $ref: "#/components/schemas/NonNegativeNumberInput"
            - description: Grams
        carbs:
          allOf:
            - $ref: "#/components/schemas/NonNegativeNumberInput"
            - description: Grams
        fat:
          allOf:
            - $ref: "#/components/schemas/NonNegativeNumberInput"
            - description: Grams; defaults to zero when omitted
        saturatedFat:
          allOf:
            - $ref: "#/components/schemas/NonNegativeNumberInput"
            - description: Grams
        transFat:
          allOf:
            - $ref: "#/components/schemas/NonNegativeNumberInput"
            - description: Grams
        fiber:
          allOf:
            - $ref: "#/components/schemas/NonNegativeNumberInput"
            - description: Grams; defaults to zero when omitted
        sodium:
          allOf:
            - $ref: "#/components/schemas/NonNegativeNumberInput"
            - description: Milligrams
        potassium:
          allOf:
            - $ref: "#/components/schemas/NonNegativeNumberInput"
            - description: Milligrams; defaults to zero when omitted
        calcium:
          allOf:
            - $ref: "#/components/schemas/NonNegativeNumberInput"
            - description: Milligrams; defaults to zero when omitted
        iron:
          allOf:
            - $ref: "#/components/schemas/NonNegativeNumberInput"
            - description: Milligrams; defaults to zero when omitted
        vitaminC:
          allOf:
            - $ref: "#/components/schemas/NonNegativeNumberInput"
            - description: Milligrams; defaults to zero when omitted
        vitaminD:
          allOf:
            - $ref: "#/components/schemas/NonNegativeNumberInput"
            - description: Micrograms; defaults to zero when omitted
        source:
          type: string
          enum: [manual, usda]
          default: manual
        date:
          $ref: "#/components/schemas/DateString"
    FoodEntryUpdate:
      type: object
      description: Any subset of food entry fields. At least one field is required.
      minProperties: 1
      properties:
        foodName:
          type: string
          minLength: 1
          maxLength: 120
        servingSize:
          type: string
          minLength: 1
          maxLength: 80
        mealType:
          $ref: "#/components/schemas/MealType"
        calories:
          $ref: "#/components/schemas/NonNegativeNumberInput"
        protein:
          $ref: "#/components/schemas/NonNegativeNumberInput"
        carbs:
          $ref: "#/components/schemas/NonNegativeNumberInput"
        fat:
          $ref: "#/components/schemas/NonNegativeNumberInput"
        saturatedFat:
          $ref: "#/components/schemas/NonNegativeNumberInput"
        transFat:
          $ref: "#/components/schemas/NonNegativeNumberInput"
        fiber:
          $ref: "#/components/schemas/NonNegativeNumberInput"
        sodium:
          $ref: "#/components/schemas/NonNegativeNumberInput"
        potassium:
          $ref: "#/components/schemas/NonNegativeNumberInput"
        calcium:
          $ref: "#/components/schemas/NonNegativeNumberInput"
        iron:
          $ref: "#/components/schemas/NonNegativeNumberInput"
        vitaminC:
          $ref: "#/components/schemas/NonNegativeNumberInput"
        vitaminD:
          $ref: "#/components/schemas/NonNegativeNumberInput"
        source:
          type: string
          enum: [manual, usda]
        date:
          $ref: "#/components/schemas/DateString"
    QuickAddFoodRequest:
      type: object
      required: [date]
      properties:
        date:
          $ref: "#/components/schemas/DateString"
        mealType:
          $ref: "#/components/schemas/MealType"
    FoodEntryResponse:
      type: object
      required: [foodEntry]
      properties:
        foodEntry:
          $ref: "#/components/schemas/FoodEntry"
    FoodEntryListResponse:
      type: object
      required: [foodEntries]
      properties:
        foodEntries:
          type: array
          items:
            $ref: "#/components/schemas/FoodEntry"
    AiFoodLogRequest:
      type: object
      additionalProperties: false
      required: [text, date]
      properties:
        text:
          type: string
          minLength: 2
          maxLength: 500
          example: I had one medium banana and a rice cake for lunch
        date:
          $ref: "#/components/schemas/DateString"
        mealType:
          allOf:
            - $ref: "#/components/schemas/MealType"
            - description: Used only when the description does not name a meal
          default: Snack
    AiFoodLogResponse:
      type: object
      required: [message, estimated, model, foodEntries]
      properties:
        message:
          type: string
          example: 2 estimated food entries logged.
        estimated:
          type: boolean
          enum: [true]
        model:
          type: string
          example: gpt-5.4-nano
        foodEntries:
          type: array
          minItems: 1
          maxItems: 8
          items:
            $ref: "#/components/schemas/FoodEntry"
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
        stack:
          type: string
          description: Stack trace, returned only outside production